Pin library versions, fix pairplot
I did not get to the root of the pairplot issue (as in, why it previously worked). The labels column used for the hue option was boolean, so was still being included in the plot (because it could be cast as a numeric type). By converting the bool to string, the labels column is removed from the plot, and no matrices are singular, so the plot renders fine.