Fix UMAP problem, move imports to top
UMAP has problems with matrices with a dimension greater than 4095, I _think_ related to scipy sparse matrices. There are several issues about it on the umap-learn repo, and I could not find an earlier version that worked (including searching through some earlier versions of scipy, pynndescent, and numba).

Also moved all imports to the top of the notebook, so that it fails immediately if requirements aren't installed, added some spaces to code cells for readability, and some general tidying up.